clean up after rebase
