+++ desktop-web/index.html
@@ -16,7 +16,7 @@
 
 <script>
 const e = React.createElement;
-// Read ./main-react.js to see what components are available
+// Read ./main.js to see what components are available
 const TwinleafApp = e(App, {toWhat: 'World'}, null);
 
 ReactDOM.render(TwinleafApp, document.querySelector('#root'));
